soovii_watermark/requirements.txt
ffmpy==0.2.2
pyPdf==1.13
reportlab==3.4.0
tornado==5.0.1

